recovery: refactor delivery rate estimation
The delivery rate estimation code is fairly self-contained, so it can be
moved into its own submodule pretty easily.
